+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-27 02:56:29
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 3f0ad7717f3255463e09a760da88de972a9a7f38
+
+## [ForwardWidgets-2kuai] - 2025-08-27 02:56:29
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 4042fb34271d9883e02ecb7fb63077b6bee75385
+
+## [ForwardWidgets-pack1r] - 2025-08-27 02:56:29
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 1d8ab46a4798700ec9dc0b268796b6ac39c305de
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-26 03:02:31
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main